# internal/auth/handler.go
package auth

import (
	"auth-strategies/internal/common"
	"encoding/json"
	"errors"
	"github.com/alexedwards/scs/v2"
	"github.com/golang-jwt/jwt/v5"
	"github.com/google/uuid"
	"github.com/rs/zerolog/log"
	"net/http"
	"time"
)

const (
	success         = "Success"
	jsonParseFailed = "JSON parse failed"
	emailTaken      = "Email address already taken"
	logoutFailed    = "Logout failed"
)

type Api struct {
	s            *Service
	sessionStore *scs.SessionManager
	hmacSecret   []byte
}

func NewApi(s *Service, sessionStore *scs.SessionManager, hmacSecret []byte) *Api {
	return &Api{s, sessionStore, hmacSecret}
}

// RegisterData payload for the register request
type RegisterData struct {
	Email     string `json:"email" validate:"required" example:"johndoe@example.com"`
	Password  string `json:"password" validate:"required" example:"foobar"`
	FirstName string `json:"firstName" validate:"required" example:"John"`
	LastName  string `json:"lastName" validate:"required" example:"Doe"`
}

// LoginData payload for login
type LoginData struct {
	Email    string `json:"email" validate:"required" example:"johndoe@example.com"`
	Password string `json:"password" validate:"required" example:"foobar"`
}

type AccessToken struct {
	AccessToken string `json:"accessToken"`
}

// Register register via email and password
//
//	@Summary	register via email and password
//	@Param		request	body	RegisterData	true	"email, full name and password"
//	@Tags		auth
//	@Produce	json
//	@Success	200	{object}	common.SuccessResponse
//	@Failure	400
//	@Failure	409
//	@Failure	500
//	@Router		/auth/register [post]
func (api *Api) Register(w http.ResponseWriter, r *http.Request) {
	registerData := &RegisterData{}
	if err := json.NewDecoder(r.Body).Decode(registerData); err != nil {
		common.WriteJSON(w, http.StatusBadRequest, common.ErrorResponse{Error: jsonParseFailed})
		return
	}

	rq := &registerRq{
		email:     registerData.Email,
		password:  registerData.Password,
		firstName: registerData.FirstName,
		lastName:  registerData.LastName,
	}
	err := api.s.register(r.Context(), rq)
	if errors.Is(err, errEmailTaken) {
		common.WriteJSON(w, http.StatusConflict, common.ErrorResponse{Error: emailTaken})
	} else if err != nil {
		log.Error().Err(err).Msg("failed to register user")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
	common.WriteJSON(w, http.StatusOK, common.SuccessResponse{Status: success})
}

// Login login via email and password
//
//	@Summary	login via email and password
//	@Param		request	body	LoginData	true	"email and password"
//	@Tags		auth
//	@Produce	json
//	@Success	200	{object}	common.SuccessResponse
//	@Failure	401
//	@Failure	500
//	@Header		200			{string}	Set-Cookie	"Session cookie"
//	@Router		/auth/login	[post]
func (api *Api) Login(w http.ResponseWriter, r *http.Request) {
	id := api.loginHelper(w, r)
	if id == nil {
		return
	}

	api.sessionStore.Put(r.Context(), "user_id", id.String())
	common.WriteJSON(w, http.StatusOK, common.SuccessResponse{Status: success})
}

// Logout log the user out of the current session
//
//	@Summary	log the user out of the current session
//	@Tags		auth
//	@Produce	json
//	@Success	200	{object}	common.SuccessResponse
//	@Failure	500
//	@Router		/auth/logout	[post]
func (api *Api) Logout(w http.ResponseWriter, r *http.Request) {
	err := api.sessionStore.Destroy(r.Context())
	if err != nil {
		log.Error().Err(err).Msg("failed to destroy session")
		common.WriteJSON(w, http.StatusInternalServerError, common.ErrorResponse{Error: logoutFailed})
		return
	}
	common.WriteJSON(w, http.StatusOK, common.SuccessResponse{Status: success})
}

// LoginToken exchange email and password for an access and refresh token
//
//	@Summary	exchange email and password for an access and refresh token
//	@Param		request	body	LoginData	true	"email and password"
//	@Tags		auth
//	@Produce	json
//	@Success	200	{object}	AccessToken
//	@Failure	401
//	@Failure	500
//	@Router		/auth/token/login	[post]
func (api *Api) LoginToken(w http.ResponseWriter, r *http.Request) {
	id := api.loginHelper(w, r)
	if id == nil {
		return
	}

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
		"sub": id.String(),
		"exp": time.Now().Add(1 * time.Hour).Unix(),
		"alg": jwt.SigningMethodHS256.Alg(),
	})
	tokenString, err := token.SignedString(api.hmacSecret)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Error().Err(err).Msg("failed to sign access token")
		return
	}

	common.WriteJSON(w, http.StatusOK, AccessToken{AccessToken: tokenString})
}

func (api *Api) loginHelper(w http.ResponseWriter, r *http.Request) *uuid.UUID {
	loginData := &LoginData{}
	if err := json.NewDecoder(r.Body).Decode(loginData); err != nil {
		common.WriteJSON(w, http.StatusBadRequest, common.ErrorResponse{Error: jsonParseFailed})
		return nil
	}

	id, err := api.s.checkPassword(r.Context(), loginData.Email, loginData.Password)
	if errors.Is(err, errInvalidCredentials) {
		w.WriteHeader(http.StatusUnauthorized)
		return nil
	} else if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Error().Err(err).Msg("login failed")
		return nil
	}

	return id
}